Add missing time support to gate_functions
Added the extra time parameter in the (X_gate, Y_gate, Z_gate, had_from_rot, CNOT_from_CZ, CRX_from_CZ) functions during the 'builder < ' call, so that the these gates are created with the correct time tag. (circuit_builder.__lt__ was modified in the previous commit to allow for this.